requirements.txt
Flask
flask-cors
gunicorn
psycopg2-binary
cloudinary
qrcode
pypdf2